Channel splitting bug fixing done.
On performing the crop operation, size of editing_img shrinks while size of edited_img does not.
If split channel filter is used after this without using Apply Changes or Cancel changes, then the size of both the images are not same.
Due to this, the pixel data from edited_img could not be put into the editing_img and resulted in an exception.

To avoid this, as soon as the split channel tab is opened, editing_img is assigned copy of edited_img i.e, cancel changes functionality is applied as default.
This ensures that size of both the images are same.