Updated Executable Conversion Method
Modified the way to include the image files in AppFrame.py so as to make the final executable completely standalone.
The executable now can display button icons even if the image files are not explicitly provided in a static folder.
The executable extracts the images in the system's temporary folder which the resource_Path fucntion directly accesses to display the icons.

Deleted the extra executable folder and Merged the 2 README files for Executables in the Project's README file.